Add json file reading and publishing mode
Add support for reading and publishing the contents of the
provided JSON file directly to the the provided topic on the command line.

This is in addition to the existing file reading mode that supports a
tab-delimited file containing records having both the topic and the json
message contents to publish to the topic.

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>